requirements.txt
fastapi
deta
python-dotenv
